requirements.txt
pandas
requests
pillow
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
gspread
dotenv
